+++ include/estimator.h
@@ -65,14 +65,14 @@ namespace theia
 		// successful model estimation (and outputs model), false for failed
 		// estimation. Typically, this is a minimal set, but it is not required to be.
 		virtual bool estimateModel(const Datum& data,
-			const int *sample, 
+			const size_t *sample, 
 			std::vector<Model>* model) const = 0;
 
 		// Estimate a model from a non-minimal sampling of the data. E.g. for a line,
 		// use SVD on a set of points instead of constructing a line from two points.
 		// By default, this simply implements the minimal case.
 		virtual bool estimateModelNonminimal(const Datum& data,
-			const int *sample,
+			const size_t *sample,
 			size_t sample_number,
 			std::vector<Model>* model) const = 0;
 
@@ -144,7 +144,7 @@ namespace theia
 		// check or some other verification of the model structure.
 		virtual bool isValidModel(const Model& model,
 			const Datum& data,
-			const std::vector<int> &inliers,
+			const std::vector<size_t> &inliers,
 			const double threshold) const
 		{ return true; }
 	};
