+++ include/fundamental_estimator.h
@@ -40,7 +40,7 @@ public:
 	}
 
 	bool estimateModel(const cv::Mat& data,
-		const int *sample,
+		const size_t *sample,
 		std::vector<Model>* models) const
 	{
 		// Model calculation by the seven point algorithm
@@ -157,7 +157,7 @@ public:
 	// epipolar distance as well. 
 	bool isValidModel(const Model& model,
 		const cv::Mat& data,
-		const std::vector<int> &inliers,
+		const std::vector<size_t> &inliers,
 		const double threshold) const
 	{
 		size_t inlier_number = 0; // Number of inlier if using symmetric epipolar distance
@@ -178,7 +178,7 @@ public:
 
 	bool estimateModelNonminimal(
 		const cv::Mat& data,
-		const int *sample,
+		const size_t *sample,
 		size_t sample_number,
 		std::vector<Model>* models) const
 	{
@@ -216,7 +216,7 @@ public:
 
 	inline bool normalizePoints(
 		const cv::Mat& data, // The data points
-		const int *sample, // The points to which the model will be fit
+		const size_t *sample, // The points to which the model will be fit
 		size_t sample_number,// The number of points
 		cv::Mat &normalized_points, // The normalized point coordinates
 		Eigen::Matrix3d &T1, // The normalizing transformation in the first image
@@ -311,7 +311,7 @@ public:
 	}
 
 	inline bool solverEightPoint(const cv::Mat& data,
-		const int *sample,
+		const size_t *sample,
 		size_t sample_number,
 		std::vector<Model>* models) const
 	{
@@ -372,8 +372,8 @@ public:
 	}
 
 	inline bool solverSevenPoint(const cv::Mat& data,
-		const int *sample,
-		int sample_number,
+		const size_t *sample,
+		size_t sample_number,
 		std::vector<Model>* models) const
 	{
 		Eigen::MatrixXd coefficients(sample_number, 9);
@@ -532,8 +532,8 @@ public:
 
 	inline int all_ori_valid(const Eigen::Matrix3d &F,
 		const cv::Mat &data, 
-		const int *sample, 
-		int N) const
+		const size_t *sample,
+		size_t N) const
 	{
 		Eigen::Vector3d ec;
 		double sig, sig1;
